MDA-508 fix to 1.8.0
