+++ app/src/main/java/com/unlimint/di/component/ApplicationComponent.kt
@@ -4,8 +4,6 @@ import com.unlimint.di.module.ApplicationModule
 import com.unlimint.di.scope.ApplicationScope
 import com.unlimint.App
 import com.unlimint.view.HomeScreenFragment
-import com.unlimint.view.PaymentMethodFragment
-import com.unlimint.view.PaymentWithTokenFragment
 import dagger.BindsInstance
 import dagger.Component
 
@@ -17,8 +15,6 @@ import dagger.Component
 interface ApplicationComponent {
 
     fun inject(homeScreenFragment: HomeScreenFragment)
-    fun inject(paymentMethodFragment: PaymentMethodFragment)
-    fun inject(paymentMethodFragment: PaymentWithTokenFragment)
 
     @Component.Builder
     interface Builder {
