--- app/src/main/res/layout/fragment_payment_with_token_layout.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <com.unlimint.ui.SdkDemoEditText
-        android:id="@+id/tokenInputLayout"
-        style="@style/SdkDemoStyle.EditText"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:hintText="@string/token_input_hint_text"
-        app:imeOptions="actionNext"
-        app:labelText="@string/token_input_label_text" />
-
-    <android.widget.Button
-        android:id="@+id/buyButton"
-        style="@style/SdkDemoStyle.Button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/sdk_demo_space_x2"
-        android:layout_marginTop="@dimen/sdk_demo_space_x2"
-        android:layout_marginEnd="@dimen/sdk_demo_space_x2"
-        android:enabled="false"
-        android:text="@string/button_continue_with_token_text" />
-
-</LinearLayout>
